+++ node/resolvers/session/index.ts
@@ -2,6 +2,7 @@ import { serialize } from 'cookie'
 import { identity } from 'ramda'
 import { sessionFields } from './sessionResolver'
 import { fieldResolvers as sessionPickupResolvers } from './sessionPickup'
+import { vtexIdCookies } from '../../utils/vtexId'
 
 const VTEX_SESSION = 'vtex_session'
 
@@ -32,10 +33,10 @@ export const queries = {
    */
   getSession: async (_: any, __: any, ctx: Context) => {
     const {
-      clients: { session },
+      clients: { customSession },
       cookies,
     } = ctx
-    const { sessionData } = await session.getSession(
+    const { sessionData } = await customSession.getSession(
       cookies.get(VTEX_SESSION)!,
       ['*']
     )
@@ -50,15 +51,16 @@ interface ImpersonateArg {
 export const mutations = {
   impersonate: async (_: any, { email }: ImpersonateArg, ctx: Context) => {
     const {
-      clients: { session, checkout },
+      clients: { customSession, checkout },
       cookies,
     } = ctx
 
-    await session.updateSession(
+    await customSession.updateSession(
       IMPERSONATED_EMAIL,
       email,
       [],
-      cookies.get(VTEX_SESSION)!
+      cookies.get(VTEX_SESSION)!,
+      vtexIdCookies(ctx)
     )
 
     const orderForm = await checkout.orderForm()
@@ -88,15 +90,16 @@ export const mutations = {
 
   depersonify: async (_: any, __: any, ctx: Context) => {
     const {
-      clients: { session, checkout },
+      clients: { customSession, checkout },
       cookies,
     } = ctx
 
-    await session.updateSession(
+    await customSession.updateSession(
       IMPERSONATED_EMAIL,
       '',
       [],
-      cookies.get(VTEX_SESSION)!
+      cookies.get(VTEX_SESSION)!,
+      vtexIdCookies(ctx)
     )
 
     try {
@@ -120,15 +123,16 @@ export const mutations = {
   savePickupInSession: async (_: any, args: SavePickupArgs, ctx: Context) => {
     const { address, name } = args
     const {
-      clients: { session },
+      clients: { customSession },
       cookies,
     } = ctx
 
-    await session.updateSession(
+    await customSession.updateSession(
       'favoritePickup',
       { address: convertCheckoutAddressToProfile(address), name },
       [],
-      cookies.get(VTEX_SESSION)!
+      cookies.get(VTEX_SESSION)!,
+      vtexIdCookies(ctx)
     )
 
     return queries.getSession({}, {}, ctx)