+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Fix unexpected deletion of cookies from session
+
+### Added
+- Temporary custom session client with update session fix.
 
 ## [2.115.0] - 2020-02-13
 ### Added